backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "motor>=3.6.0",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.31.1",
]
